v0.2.66
All references to *secid*s have been removed. This will allow the 
renaming of *ahid*s to *secid*s to bring it back into consistency with 
the OM software specification and the NESTS documentation.

The implementation of direct *account*-to-*account* payments will still 
be possible, but *accounts* will all belong directly to the *primid* 
(functions to be restored) rather than indirectly via *secid*s.